# hclust() enhancements