--- a/src/hclust.jl
+++ b/src/hclust.jl
@@ -16,7 +16,7 @@ end
 function assertdistancematrix(d::AbstractMatrix)
     nr, nc = size(d)
     nr == nc || throw(DimensionMismatch("Distance matrix should be square."))
-    issymmetric(d) || error("Distance matrix should be symmetric.")
+    issymmetric(d) || throw(ArgumentError("Distance matrix should be symmetric."))
 end
 
 ## This seems to work like R's implementation, but it is extremely inefficient