… benchmark/benchmarks.jl
… benchmark/print_table.jl
… doc/source/hclust.rst
… src/types.jl
… test/data/hclust_generated_examples.jl.gz
… test/hclust-gen-examples.R
… test/hclust-generated-examples.jl
… test/hclust.jl