It started as a one line fix to reduce the number of allocations.
Actually, despite the big diff, the code of @davidavdav is quite efficient.
The size is due to factoring out the common tree maintenance logic into separate methods (that should simplify adding new hclustering methods in future).

### Notable changes:
* `HclustMerges` and `HclustTrees` classes that maintain the history of tree merges and the list of active trees
* reduced N of allocations and memory usage (in particular, by avoiding allocating 1-element vectors for the leaf nodes)
* more efficient leaves ordering (`hclust_perm()`) and `cutree()`
* the cluster numbers returned by `cutree()` match the R version
* `hclust_n3()` and `cutree()` covered by tests
* added rudimenary benchmark scripts

## Benchmarks (the script is in `benchmark` folder):

### `hclust()` for 10000 points, `:single` method
master
```
julia> hclust_old["hclust"]["10000"]["single"]
BenchmarkTools.Trial: 
  memory estimate:  975.62 MiB
  allocs estimate:  238837
  --------------
  minimum time:     6.239 s (3.15% GC)
  median time:      6.239 s (3.15% GC)
  mean time:        6.239 s (3.15% GC)
  maximum time:     6.239 s (3.15% GC)
  --------------
  samples:          1
  evals/sample:     1
```

PR
```
julia> hclust_new["hclust"]["10000"]["single"]
BenchmarkTools.Trial: 
  memory estimate:  855.21 MiB
  allocs estimate:  21101
  --------------
  minimum time:     5.411 s (2.24% GC)
  median time:      5.411 s (2.24% GC)
  mean time:        5.411 s (2.24% GC)
  maximum time:     5.411 s (2.24% GC)
  --------------
  samples:          1
  evals/sample:     1
```

### `hclust()` for 10000 points, `:average` method
master
```
julia> hclust_old["hclust"]["10000"]["average"]
BenchmarkTools.Trial: 
  memory estimate:  9.74 MiB
  allocs estimate:  147902
  --------------
  minimum time:     55.444 s (0.00% GC)
  median time:      55.444 s (0.00% GC)
  mean time:        55.444 s (0.00% GC)
  maximum time:     55.444 s (0.00% GC)
  --------------
  samples:          1
  evals/sample:     1
```

PR
```
julia> hclust_new["hclust"]["10000"]["average"]
BenchmarkTools.Trial: 
  memory estimate:  5.18 MiB
  allocs estimate:  33134
  --------------
  minimum time:     36.340 s (0.00% GC)
  median time:      36.340 s (0.00% GC)
  mean time:        36.340 s (0.00% GC)
  maximum time:     36.340 s (0.00% GC)
  --------------
  samples:          1
  evals/sample:     1
```

### `cutree()` for `Hclust` of 10000 points into 1000 clusters
master
```
julia> hclust_old["cutree"]["(10000, 1000)"]
BenchmarkTools.Trial: 
  memory estimate:  88.64 MiB
  allocs estimate:  105318
  --------------
  minimum time:     24.714 ms (39.07% GC)
  median time:      27.419 ms (42.49% GC)
  mean time:        31.458 ms (48.47% GC)
  maximum time:     113.813 ms (85.54% GC)
  --------------
  samples:          161
  evals/sample:     1
```

This PR
```
julia> hclust_new["cutree"]["(10000, 1000)"]
BenchmarkTools.Trial: 
  memory estimate:  42.86 MiB
  allocs estimate:  12374
  --------------
  minimum time:     14.110 ms (7.31% GC)
  median time:      15.554 ms (12.31% GC)
  mean time:        16.288 ms (12.05% GC)
  maximum time:     31.050 ms (8.68% GC)
  --------------
  samples:          307
  evals/sample:     1
```

cc @davidavdav 

Maybe it also helps #93