@alyst 2018-08-20
So I've rebased it and updated to v0.7 plus implemented fixes and tests for #109.
Here's the latest benchmarks:

| func   | param         | method   | time_old   | time_new   | gcratio_old | gcratio_new | allocs_old | allocs_new | memory_old | memory_new | samples_old | samples_new |
|--------|---------------|----------|------------|------------|-------------|-------------|------------|------------|------------|------------|---------------|---------------|
| cutree | (100, 10)     | missing  | 21.381 μs  | 6.337 μs   | 0.00%       | 0.00%       | 790        | 71         | 75.94 KiB  | 11.61 KiB  | 10000         | 10000         |
| cutree | (10000, 1000) | missing  | 19.788 ms  | 44.891 ms  | 42.36%      | 78.65%      | 78297      | 13029      | 70.10 MiB  | 53.58 MiB  | 225           | 2             |
| cutree | (10, 3)       | missing  | 1.805 μs   | 838.000 ns | 0.00%       | 0.00%       | 71         | 15         | 5.67 KiB   | 1.16 KiB   | 10000         | 10000         |
| cutree | (1000, 100)   | missing  | 302.779 μs | 174.909 μs | 0.00%       | 0.00%       | 7836       | 1004       | 1.36 MiB   | 514.67 KiB | 10000         | 255           |
| hclust | 100           | average  | 713.783 μs | 251.955 μs | 0.00%       | 0.00%       | 850        | 203        | 77.39 KiB  | 33.02 KiB  | 6843          | 10000         |
| hclust | 100           | single   | 267.058 μs | 74.105 μs  | 0.00%       | 0.00%       | 1641       | 73         | 187.78 KiB | 96.66 KiB  | 10000         | 10000         |
| hclust | 100           | complete | 927.169 μs | 255.127 μs | 0.00%       | 0.00%       | 770        | 220        | 69.39 KiB  | 33.75 KiB  | 5305          | 10000         |
| hclust | 10            | average  | 8.264 μs   | 5.204 μs   | 0.00%       | 0.00%       | 97         | 44         | 7.34 KiB   | 3.88 KiB   | 10000         | 10000         |
| hclust | 10            | complete | 8.634 μs   | 5.050 μs   | 0.00%       | 0.00%       | 97         | 48         | 7.31 KiB   | 4.11 KiB   | 10000         | 10000         |
| hclust | 10            | single   | 17.332 μs  | 3.354 μs   | 0.00%       | 0.00%       | 147        | 29         | 8.80 KiB   | 3.30 KiB   | 10000         | 10000         |
| hclust | 10000         | complete | 57.037 s   | 10.812 s   | 0.00%       | 0.00%       | 76724      | 35329      | 7.58 MiB   | 4.74 MiB   | 1             | 1             |
| hclust | 10000         | average  | 54.697 s   | 11.747 s   | 0.00%       | 0.00%       | 88237      | 33537      | 8.16 MiB   | 4.75 MiB   | 1             | 1             |
| hclust | 10000         | single   | 3.544 s    | 2.755 s    | 4.40%       | 4.99%       | 173901     | 12192      | 976.94 MiB | 897.92 MiB | 2             | 2             |
| hclust | 1000          | complete | 174.312 ms | 53.446 ms  | 0.00%       | 0.00%       | 7285       | 2740       | 711.88 KiB | 391.44 KiB | 29            | 85            |
| hclust | 1000          | average  | 136.956 ms | 56.907 ms  | 0.00%       | 0.00%       | 8386       | 2597       | 826.61 KiB | 391.72 KiB | 36            | 81            |
| hclust | 1000          | single   | 12.352 ms  | 11.281 ms  | 0.00%       | 0.00%       | 16179      | 919        | 10.46 MiB  | 8.83 MiB   | 395           | 301           |

I think there was some glitch in `cutree (10000, 1000)`, which would go away in the next try. In the other tests the new code shows improvements.
As before, the results are compared with R's `hclust()` for a few hundred random datasets, so the algorithms should be correct.

It would be nice if somebody reviews.

This PR introduces some helper functions to facilitate the construction of the hierarchical clustering. My plan is to implement Ward clustering (or a generic Lance-Williams one) on top of it in the PR to follow. There was already #94, but it had license issues. The new code will be written from scratch, so it should have no license issues.