@alyst 2018-08-23
This PR is already big, so I've added Ward (`:ward` == R's `Ward.D2` and `:ward_presquared` == R's `Ward.D1`) here. Ward and any other Lance-Williams metrics (including `:average` and `:complete`) are supported by the new `hclust_nn_lw()` (in the end it's not much different from `hclust2()`, but much faster since it doesn't recalculate the metric from point distances for every comparison). I've made `hclust_nn_lw()` the default method. It uses more RAM (it has to maintain its copy of distance matrix), but that should be fine.

I've renamed some `Hclust` fields (`height` => `heights` etc) and implemented support for the old names during deprecation period via `getproperty()`.

I would merge it in a few days, if nobody objects. Still happy to see the reviews. :)

The new benchmark (`:single` got 10% slower since the last benchmark, because I made it share some common routines with the `hclust_nn_lw()`  (neighbor search, matrix updates), but it should not matter much in practice).

| func   | param         | linkage  | time_old   | time_new   | gcratio_old | gcratio_new | allocs_old | allocs_new | memory_old | memory_new | n_old  | n_new |
|--------|---------------|----------|------------|------------|-------------|-------------|------------|------------|------------|------------|--------|-------|
| cutree | (10, 3)       | single   | 1.805 μs   | 11.412 μs  | 0.00%       | 0.00%       | 71         | 15         | 5.67 KiB   | 1.16 KiB   | 10000  | 902   |
| cutree | (100, 10)     | single   | 21.381 μs  | 21.040 μs  | 0.00%       | 0.00%       | 790        | 77         | 75.94 KiB  | 11.36 KiB  | 10000  | 938   |
| cutree | (1000, 100)   | single   | 302.779 μs | 211.556 μs | 0.00%       | 0.00%       | 7836       | 1000       | 1.36 MiB   | 464.11 KiB | 10000  | 499   |
| cutree | (10000, 1000) | single   | 19.788 ms  | 16.917 ms  | 42.36%      | 54.46%      | 78297      | 13037      | 70.10 MiB  | 51.24 MiB  | 225    | 11    |
| hclust | 100           | average  | 713.783 μs | 77.113 μs  | 0.00%       | 0.00%       | 850        | 301        | 77.39 KiB  | 112.53 KiB | 6843   | 10000 |
| hclust | 100           | complete | 927.169 μs | 94.241 μs  | 0.00%       | 0.00%       | 770        | 318        | 69.39 KiB  | 113.42 KiB | 5305   | 10000 |
| hclust | 100           | single   | 267.058 μs | 78.902 μs  | 0.00%       | 0.00%       | 1641       | 74         | 187.78 KiB | 96.53 KiB  | 10000  | 10000 |
| hclust | 10            | complete | 8.634 μs   | 4.754 μs   | 0.00%       | 0.00%       | 97         | 56         | 7.31 KiB   | 4.89 KiB   | 10000  | 10000 |
| hclust | 10            | average  | 8.264 μs   | 4.181 μs   | 0.00%       | 0.00%       | 97         | 52         | 7.34 KiB   | 4.70 KiB   | 10000  | 10000 |
| hclust | 10            | single   | 17.332 μs  | 2.783 μs   | 0.00%       | 0.00%       | 147        | 29         | 8.80 KiB   | 3.31 KiB   | 10000  | 10000 |
| hclust | 10000         | complete | 57.037 s   | 2.520 s    | 0.00%       | 0.35%       | 76724      | 45128      | 7.58 MiB   | 767.69 MiB | 1      | 14    |
| hclust | 10000         | single   | 3.544 s    | 3.104 s    | 4.40%       | 4.35%       | 173901     | 11997      | 976.94 MiB | 895.62 MiB | 2      | 12    |
| hclust | 10000         | average  | 54.697 s   | 2.341 s    | 0.00%       | 0.38%       | 88237      | 43384      | 8.16 MiB   | 767.80 MiB | 1      | 15    |
| hclust | 1000          | complete | 174.312 ms | 13.866 ms  | 0.00%       | 0.00%       | 7285       | 3717       | 711.88 KiB | 8.02 MiB   | 29     | 2586  |
| hclust | 1000          | average  | 136.956 ms | 11.137 ms  | 0.00%       | 0.00%       | 8386       | 3550       | 826.61 KiB | 8.02 MiB   | 36     | 2929  |
| hclust | 1000          | single   | 12.352 ms  | 12.805 ms  | 0.00%       | 0.00%       | 16179      | 911        | 10.46 MiB  | 8.80 MiB   | 395    | 2493  |

cc @ararslan (As you were the most active in the repository recently :) )